Housekeeping -change packages
I ran npm update on our modules in order to make the process go as fast as possible. This changed the root package.json by updating mongodb, npm, and the overall version number of the build.
Because as of npm v. 5.1.0 the package overwrites the packagelock if there's a newer compatable version. Both package-locks were changed to use lockfileVersion 1 for some reason I couldn't find.

Deleted yarn.lock and added it to gitignore since it is a user generated file.

This version of modules works with no errors / warning where there were previously.

Note: If anything breaks in production the changed package-lock s may be at fault, in which case we should roll this back. This version was able to start much faster and install quicker during testing, though.